bugfix: built_at doesn't give us enough precision for sorting and can lead to position hijacking.
You could end up with newly uploaded gems being shown several places down in the dashboard or atom feed for latest gems.

created_at may not accurately show when the gem was built on the persons computer v.s. when it was uploaded, but at least it will order things in the way most people would expect to see them (newest uploaded)

Includes test which details a little more what order they should be in and why.

Signed-off-by: Nick Quaranto <nick@quaran.to>