Use the find_gems method defined in the ApplicationController
The find_rubygems method defined in this class does the same thing as
the find_gems method defined in the ApplicationContoller except that it
doesn't have any error handling. This leads to an Internal Server Error
(instead of 404) when a gem/version combination is not found.
For example: https://rubygems.org/gems/rails/versions/0.0.0