Adds Vagrant setup for development
Vagrant setup based on the Development Setup on the wiki.

The setup contains:
- Ruby 1.9.3 managed with chruby and built with ruby-build
- PostgreSQL
- Redis

Port 3000 is forwarded to your localhost, so when running 'rails s' on
the vm you can visit http://localhost:3000 on your own machine to visit
the app.

When on Linux or Mac OS X NFS is used to share the source directory
between your machine and the Vagrant box. To use NFS the machine must
have an IP address, it is given 192.168.33.100 (also on Windows).

If you create the directory .apt-cache/partial in your source directory
the aptitude cache will be shared there between your machine and the
Vagrant box. This means you won't have to redownload all packages from
apt between box rebuilds.