Disallow name case changes when there are indexed versions
This makes the lookup case insensitive still, which protects against
multiple users having gems like RAILS or nokogirI. But it fixes an edge
case where the owner might want to change the gem case. We don't allow
changing it when there are indexed versions because doing so screws up
the index matching the filenames of the gems on disk.