Speed up the dashboard feature test
I can't think of a very good reason to call `visit` 2,009 times for this
test. Changing these to more sane numbers took this test from ~8 seconds
to 0.5 seconds.

If this spec was trying to test the formatting of the numbers on the
dashboard then a view spec would much more simply and efficiently cover
that. I don't think that's a particularly valuable test so I didn't
implement it here.