Keep a consistent sans-serif font
This pull request is very font-nerdy :books: but here it goes.

One line of CSS specifies that Arial should be used over Helvetica, when possible. The other line of CSS specifies that Helvetica should be used over Arial, when possible. For visual consistency, the two lines should be the same.

And indeed, there is no need to specify either of those options, since `font-family: sans-serif` [already makes that choice for you](http://css-tricks.com/sans-serif), defaulting to Helvetica on Mac (where it's available) and to Arial on Windows.