Fix the size of Clippy because it prevents Copy.
A commit made in #433 / #440 is broken in recent versions of Chrome which do not like Objects to extend past their bounds, the oversizing of the area causes Chrome to trip out (at least on Linux, I assume on Windows and OS X too) because it extends into the text area and thus prevents any manual copy and paste.  This could also be broken by other changes and might not necessarily be related to Chrome at all.

The size of 24px was manually selected after adjusting the maximum size in inspector until I could copy.