Remove not allowed chars from name before use as url.
Unfortunately, we have some gems that contains some chars that are not
resolved as the URL pattern we have on the routes, so we cannot try to
create a link out of them. Ideally we should not save those names, as we
have a validation on the Rubygem model, however we still have some bad
data in the db, so when converting to_param, we need to clean them up before.