Remove Hostess Sinatra app from production
As far as we can tell, this routes are not done in ngnix. So this
middleware will only be necessary in development.
In this commit I am only moving this out of production so we can make
sure everything still works, and I will follow up in a next commit to
make the Hostess middleware a way simple class.