Improved dutch translations
In some places the dutch translations were a bit off, or just did not sound
correctly. This commit translated some more content, and also fixes some of
those tags that did not sound right.

I also re-ordered the nl.yml a bit so all the non-namespaced keys are at the top
of the file.

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>