Gaug.es should use async defer, not script injection
Thanks to the browser preloader, using "async" and "defer" attributes will always
be faster in modern browsers. When using script injection, the preloader cannot
start downloading the resource until it evaluates the script (which is currently
at the bottom of the body, after a very large blocking script eval from Typekit).