Query only runtime or dev reverse_dependencies
This adds a new query parameter (`only`) to the Reverse Dependencies API endpoint in order to return exclusively gems that declare the provided gem as a runtime or development dependency.

It's often much more important to know which gems use your gem at runtime than all the gems that use your gem in development. In this case
you would pass `?only=runtime`.

If you're only interested in development dependencies, you can use
`?only=development`.

It may make sense to create a whole separate endpoint for this but it seemed simple enough to add a query param for this.