Fix N+1 queries on activity api
latest and just_updated endpoints need to load information from a few
other tables, which was causing some big N+1 queries on those endpoints.

Also, improve the query for `latest`, as that didnt have a limit on the
inner query, and eliminate the `latest` flag check, as we already check
if thats the only version for that rubygem.