Add version navigation links
This is a really basic feature I find myself needing often when looking for
runtime dependency changes in a gem.

![Demo](https://s3.amazonaws.com/f.cl.ly/items/030o0t280F2p3N3T403b/Screen%20Recording%202016-02-19%20at%2012.03%20PM.gif?v=7fc83002)

It adds two simple methods `Version#previous` and `Versions#next` which just use
`self.position - 1` and `self.position + 1`. If adding the links to the UI seems
unnecessary we could even hook up keyboard shortcuts to make this a power-user
feature only but I'll admit I prefer discoverable features.

If this feature makes sense, let me know and I'll add tests. :-)

The demo doesn't show the dependencies below but that's mostly the advantage