Use Postgresql to push/pop log files to be processed.
SQS delivers the messages as at-least-one, so we need to make sure our
job is idempotent. Previous implentation was rely on Redis for that, however
there was a TTL on it, which could cause problems if a message is delivered again
after the TTL.

Besides we would like to remove Redis from our stack, for a few other
resons.