Add a width to the footer.
As it stood with the old CSS if there was no content then the footer
would only fill half the page.  This fixes it so that if whether there
is content or not then the footer will fill the entire width of the
page.  This problem was noticed in Chrome/Firefox on Linux x86-64.