Remove versions N+1
There was a N+1 on versions table on the /gems endpoint. That was
happening because we show the version and description for the gem, and
we need to get that from the latest version.

Now we have a Rails association for the latest version, so we can eager
load that.