Prevent stats page error when no downloads
If the stats page was visited while there were no downloads in the
database, a NoMethodError would be thrown due to
`@most_downloaded.first` being nil. This commit prevents this by
checking it is available before accessing it.

While this issue wouldn't occur in production it could in development
after following the standard instructions for setting up a development
environment in CONTRIBUTING.md.