Changes to fix broken tests
Reusing params in ES suggestion view was showing error for unsafe use of
params.
`count` takes `conditions` as argument , but we don't need even that.
Sytax of content_type changed.
Rails now uses async queue_adapter which uses a different thread to
work off jobs. We can't let that happen in our integration test because
we are relying on the mail being delievered and doing that in
a different thread leads to unpredictable delay.