Use a seperate variable for user edit form
When `current_user.update_attributes` fails, the invalid
values are not stored in DB. However, it changes local fields of
`current_user` so that form can be rendered with errors and invalid
data in form fields.
We use `current_user` in edit page to render elements beside the
form (for ex: navigation bar user link). Failed update to `current_user`
changed those unrealted parts of page.