Base queries on versions instead of gems
The popular action doesn’t work because of the issues discussed in
66765d45a4472fb458bf9fa47a83f092941a5016#commitcomment-19480753

I’d be curious if anyone has a way around that would allow me to
sort versions by downloads for the popular action.

I could simply query gems first to get the 100 most downloaded and then
fetch recent versions for those.