Change the default test queue_adapter to :test
Rails 5 uses async queue_adapter which uses a different thread to
work off jobs. We can't let that happen in our integration test because
we are relying on the mail being delievered and doing that in
a different thread leads to unpredictable delay.
Using `:test` facilitites use of `perform_enqueued_jobs` when needed.