Fix assumptions made in yank test
Tests were bases on assumption that user gets signed in when yank
request is made. now it needs to done explicitly.
test for "yanked gem entirely then someone else pushes a new version"
was wrong, it was passing because it was finding handle on nav bar of
signed in user.