Add Fastly insights.js for performance monitoring
https://github.com/fastly/insights.js

RubyGems is hosted on Fastly.  This script allows Fastly to monitor the
network performance of rubygems.org pages which are served through
Fastly.

This script is GDPR compliant and does not collect any personally
identifiable information.

This script collects information about HTTP and HTTPS network
transactions, including: network routing, performance timing, and
equipment characteristics. Measurements are recorded to analyze the
performance of the Fastly network and overall state of the internet.

See the above link for further details on this script.