Simplify and speed up rack-attack integration tests
Instead of running hundreds of requests here, we can put the cache in a
condition where rack-attack will believe that the requests have happened and
then block the subsequent request made in the test.

This should reduce intermittent failures, and also speed up the test.

The setup block populates the cache for a particular IP address to simulate too
many requests, and each test supplies that IP address in the request that is
made, which ensures that the pre-populated cache is the one which is checked.
This mirrors real world behavior of the rack-attack prevention mechanism.