Simplify the reverse_dependencies index view
This was sort of confusing with a collection being passed to a partial which was
not actually the thing being collection rendered ... but was just passing the
collection on to another render after building a search form.

As a side effect of this, we are now explicitly passing instance variables to
partials where needed, instead of assuming their existence.