Add webhook for yanked versions
Adds a `yanked` attribute to the payload serialized in webhooks for a given
version, and adds a trigger to send webhooks when a gem is yanked. The webhook
sent is the same as the one sent when a new version is pushed (except the new
yanked attribute will be true).