Extract dynamic_form's error_messages_for to avoid a wrong i18n scope
Closes #1834

The problem occurs because we use `form.error_messages` in the form and
dynamic_form adds `activerecord.errors.template.{header/body}` i18n key
when it has error. The problem is that `rails-i18n` gem doesn't have
this key. It has `errors.template.{header.body}` instead. To fix this,
we did extract the method to our helper and removed this wrong scope.

We also have removed the dynamic_form dependency with this change.