For gem edit pages, build a new linkset if gem is missing one
We have a few places where we work around the issue of a gem not having
an associated linkset record, so this at least is in that genre of
change.

This was reported to us via the support site; something about that
user's gemspec is funky, and a linkset isn't being created. The unusual
thing is that this user attempted to edit via the web UI, and I wouldn't
be surprised if this is a low-traffic action, which would explain why we
wouldn't have seen this error crop up previously.

http://help.rubygems.org/discussions/problems/34216-unable-to-edit-newly-published-gem-server-error