Create seperate resources for adoption request
This was needed because we should save owner_id with approvals and
timestamp.
It will still have null value for adoption requests where status is
opened
but it is not as bad as null values with adoptions records created
by owners(previously).

I had started with one model to keep my change as small as possible.