Use build option instead of method override in test gem helper
The `Gem::Package.build` method takes an argument which allows the validation
step to be skipped, but it was not in use here. This helper skips the validation
to suppress noisy output about a (fake) gemspec defined in the test helper. We
were previously suppressing the output via a method override, this change
updates to use an option instead.