Reorder versions only when indexed column changes
Periously, versions were getting reordered for any change to version
record and causing performance issue where gems have a large set of
released versions.
We need to reorder versions only when a new version is created or when
a version is yanked (indexed column changes).
This reduces versions reordering in gem push from thrice to once and
in gem yank from twice to once.