Small refactoring of #link_to_github logic
I was writing up a script for a project on code metrics and spotted this
method as one of the more complicated ones in the app. It wasn't /bad/
but I thought it could be a bit cleaner. This change drops its flog
score from 38 to 22.7, but YMMV on if you think this is cleaner or not.