Move Hostess middleware to lib
Fixes in rails 6 update: A copy of Hostess has been removed from the
module tree but is still active! (ArgumentError) in development.rb

`app` folder is autoloaded in development and causes this error. It is
adviced to not put middleware (not reloadable) in app folder.

require_relative is used as `lib` is not present in load path during
initialization.