Add PROFILE env variable dev switch for perf work
Locally, we may want to work on improving the performance of Rubygems.org.
To do this, we need to work with production-like behavior, so that 
we can reproduce production issues and be reasonably certain that improvements
we see in development will carry over to production.

To do this, we need to change a few important settings:
* Disable code reloading
* Force the use of precompiled assets
* Skip checks like the migration middleware 
* Log to stdout, not to disk, and log at a higher level 

Adding this will make it easier to contribute high-quality performance 
improvements to Rubygems.org.