Disable typo protection for now
This is a good idea, but it's blocking a lot of legitimate gems, and
it's too hard for our limited support time to go to manually allowing
gems through.

A Slack feed of gems that look like they might be typos to allow humans
to review would be a more practical way to implement this sort of idea.
Better yet, a Slack message with a single button to yank the gem and/or
the user who pushed the gem. :looks hopefully in @colby-swandale's
direction: