Use just the dependency name as a link
Because the gem name currently includes the version restrictions as part of the link, there is not possible to copy just the version restrictions. It usually needs to copy a lot more. Since I am copying the version restrictions from Rubygems.org relatively often, this become annoying to me over the years.

So this is proposal to fix this, by not including the versions into the link. I think that just the gem dependency name should be enough, since it is enough for the versions on the left side.

However, I have not tested this at all, because I don't have rubygems.org development environment installed. So if it does not really work, please take is as a suggestion a fix it appropriately. Thx