Update GemTypo to use the -/_ variation detection
This update changes the method used by GemTypo from the Levenshtein distance
comparison with popular Gems with 10 million or more downloads to a simpler
check that compares the proposed Gem with all of its - and _ characters removed
to all existing Gems with their - and _ characters removed.

This comparison is performed completely within PostgreSQL and an index is added
to ensure that a table scan is not neccesary. This means that this typo protection
can be enforced againstn any other already published Gem rather than just popular
ones with 10 million or more downloads.

My April 2020 analysis of the typo squatting attack in February of 2020 found that
none of the target Gems were protected at the time and the vast majority of the
malicious Gem names would have been prevented by this check if it had been in place
at the time. Please see https://speakerdeck.com/rietta/securing-the-open-source-software-supply-chain
slides 22 and 23 for the publication of my findings at this time.

Fix same name error

This update fixes the error where a gem that is being upgraded was
considered a typo squat. Additionally, I simplified the protected_typo?
method a bit and ensured that only one query was neccesary rather than
two without overfetching data.

Refactor GemTypo tests

Updates to index based on internal code review and stats gathering

Fix broken test

Allow update gem pre-existing Gem that is a typo variant of another

For existing data, where there may be pre-existing typo variants such that
awesome_gem and awesome-gem both already exist. Then skip typo checking if
the current gem name matches either. This does not reduce security in that
other validations already prevent a user from publishing a Gem that has the
exact same name as a Gem already published by another user.

Consider yanked Gems when determining

This update considers the yanked_at status for potential exact name
matches such that a Gem that has at least one not-yanked version would
be considered an exact match. However, a Gem that has all versions yanked
and is also a typo squat of another Gem would be determined to be a
typo squat.

Update GemTypo to use existing case insensitive name index

This update removes the unnecsary downcasing of the target Gem name
and uses the existing UPPER(name) index on the rubygems table to check
for exact name matches.