Delete params key from params
`params` is an arguement of rails method `url_for` and it expects
hash as value. kaminari should be filtering this with `PARAM_KEY_EXCEPT_LIST`
as `params` is supported by `url_for` for internal use (not for end user
to control). However, they don't seem to agree.

Fixes:
```
TypeError: no implicit conversion of String into Hash

[GEM_ROOT]/gems/actionpack-6.0.2.2/lib/action_dispatch/routing/route_set.rb:822
:in `merge!`
[GEM_ROOT]/gems/actionpack-6.0.2.2/lib/action_dispatch/routing/route_set.rb:822
:in `url_for`
[GEM_ROOT]/gems/actionpack-6.0.2.2/lib/action_dispatch/routing/url_for.rb:180
:in `full_url_for`
[GEM_ROOT]/gems/actionpack-6.0.2.2/lib/action_dispatch/routing/url_for.rb:170
:in `url_for`
[GEM_ROOT]/gems/actionview-6.0.2.2/lib/action_view/routing_url_for.rb:89
:in `url_for`
[GEM_ROOT]/gems/kaminari-core-1.2.0/lib/kaminari/helpers/tags.rb:39 :in
`page_url_for`
```

PARAM_KEY_EXCEPT_LIST is removed as the patch not exists in kaminari.