Enable sending mail using Parameterized::DeliveryJob
Fixes deprecations:
```
DEPRECATION WARNING: Sending mail with DeliveryJob and
Parameterized::DeliveryJob is deprecated and will be removed in Rails
6.1. Please use MailDeliveryJob instead. (called from call at
/home/travis/build/rubygems/rubygems.org/lib/clearance_backdoor.rb:9)
```
Change is not backward compatible. This could be deployed during off
peak hours, when no mail is pending in queue.
If some job does fail, we can queue it again with same args manually.