Remove unreachable branch in api/v1/owners/gem endpoing
The gem method is currently using User.find_by_slug! to find the user from
the :handle params. #find_by_slug will trigger a AR::NotFound exception if
is not able to find by id nor handle. Since is an exception the now-removed
branch is not reacheable, ergo not needed.

Added some functional test to cover this case.