Prophylactic measures on /dashboard endpoint
The /dashboard endpoint is an open endpoint with the
ability to authenticate the user via and Authorization header
or the usage of api_key parameter, so the users are able to get
the dashboard via RSS/Atom . This openess also open
the opportunity to have DoS or DDoS affecting the internal
network because the above headers or parameters will be
sent as a parameter to the database, the size of these
parameter will be limited only by the limits on nginx or
unicorn before they issue a 414 response (URI Too Long).

As a prophylactic measure we should reject post with an
API key longer than expected, so no bad actors(s) would have
the chance to do more harm sending big/long request and
expect rubygems.org to have a bad time working on this
unnecessary task.

Note: I'm returning a 406 'Not Acceptable' in order to let
administrator to instrument and take actions via their
sysops tools such as HoneyBadger or similars when seeing an
elevate number of responses of this kind.