Use build-arg to set rubygems version
Fixes mismatch between expected gem version and installed gem version.
As of now installed version of gem in prod is `3.0.3`, when we expect
it to be 2.6.10.
In Dockerfile in we were copying `/usr/local/bin/gem`, which is just
a loader and doesn't decide rubygems version.
Copying `/usr/local/lib/ruby/2.6.0/rubygem` from builder would most likely
work but this seems unconventional and doesn't really save as much build time.