Update failing test on ruby 2.7
bad-date gem had date `date: 1017-12-09 00:00:00.000000000 Z`
which raises Argument error in ruby 2.6 and below.
The issue was fixed in
https://github.com/ruby/ruby/commit/e690df1f1ef4f791295448f9192d6e027400ee72
and released with 2.7, and the test no longer fails.
I have updated the date to 1017-13-09 00:00:00.000000000 Z, which
means the test and the gem is still for bad date but not same flow
(where ArgumentError was being raised)