Remove paul_revere and announcements related files
We have never used this in years and its version requirements become
a blocker every time I try to update rails version.
We generally use blog.rubygems.org or email for announcements.
We are using it on staging.rubygems.org but that too has been
redundant since we added basic auth on it.