Add note: on installing docker
### Why this PR?

Because of uniformity. All the other commands give installation instructions, with the exception of docker. I resorted to Google, only to realise that the installation instructions were contained above. Hopefully the note will direct readers to the links above. A small change, i hope it is helpful. if not, pls feel to close.

Ben