Don't redirect dependency api endpoints to https on host insecure.rubygems.org
We want to allow users to use `--source 'http://insecure.rubygems.org'`
in cli and not have it redirect to https. It can be possible that user's host
doesn't have openssl installed, which is required for https source.
We use to support this on all domains before migrating to `config.ssl = true`.