Remove format validation from api-key name
we use hostname and whoami in default key name. Limiting the format
when we aren't making users always pick the name was not feasible.

Alternatively, we can ask the user to run `gem signin` in error
message which would allow them to pick key name.
Rather misleading error message on failed gem push was:
> Name can only contain letters, numbers, dash and underscores