Add missing pt-BR translations on Signup page
This commit adds three missing translations for pt-BR on the Signup
page:
- User model name
- password previously appeared in a data breach error
- Password field label

To fix the password label translation this commit manually adds the
translation of the attribute instead of relying on the action-view auto
translation. This happens because action-view looks before for a
helpers.label.password translation before trying to use the
human_attribute_name of it and Clearance defines this translation key in
their Gem on english and we fallback to it since we don't have it
defined on other idioms.