Don't abbreviate MFA when you don't need to
MFA is complex and confusion enough without making it sound esoteric as well.

This PR spells out "multifactor authentication" the same way as it's been spelled in
existing localization files (although it should be spelled `multi-factor authentication`)
so that people who don't understand what the hell `MFA` means aren't instantly turned away
by the initialism.

Since MFA is also referred to as `Two Factor Authentication` or `2FA` by many other systems, spelling it allows
users who are at least familiar with that to pattern-match the words.

Yes, it's a bit pedantic, but we need every arrow in our quiver to get people on board with more secure authentication methods.