Disambiguate usage of `gem yank`
Doing:

```
$ gem push my.gem
Pushing gem to https://rubygems.org...
Repushing of gem versions is not allowed.
Please use `gem yank` to remove bad gem releases.
```

May be interpreted as one should yank to be able to override. However,
wether the gem was yanked or not, it is not possible to override it.

Hence we'd rather advise to push a new version, yanking is rarely the
most appropriate option.

See https://github.com/rubygems/rubygems.org/issues/2400 for a
discussion on why _unyanking_ is not possible nowadays, and the reasons
that lead to this commit.

Signed-off-by: Ulysse Buonomo <buonomo.ulysse@gmail.com>