Prevent remaining possible silent failures for ownership create
Ownerships are created in the following places

- The owners API controller (the original location of the reported issue) https://github.com/rubygems/rubygems.org/blob/fc7cc3a9879d69adb48e2d37282aea34d7f311b0/app/controllers/api/v1/owners_controller.rb#L18-L21
- The owners controller https://github.com/rubygems/rubygems.org/blob/fc7cc3a9879d69adb48e2d37282aea34d7f311b0/app/controllers/owners_controller.rb#L34-L36
- The pusher https://github.com/rubygems/rubygems.org/blob/fc7cc3a9879d69adb48e2d37282aea34d7f311b0/app/models/pusher.rb#L127-L136
- The ownership request model https://github.com/rubygems/rubygems.org/blob/fc7cc3a9879d69adb48e2d37282aea34d7f311b0/app/models/ownership_request.rb#L16-L21

In the first two, the original `.create` has already been replaced by a
`.save` which has its return value checked. This fixes the initial issue
in #2234

In the last two, the user for the ownership is already persisted and an
ownership will only fail validation if a duplicate ownership  exists.
This should be a rare occurrence. However, I went ahead and added some
extra safety just in case.