replace all instances of whitelist with allowlist
This commit replaces all instances of whitelist with allowlist. None of
this terminology is user facing or exposed via the API therefore the
changes are backwards compatible. This feels like a small change to make
the codebase more welcoming to contributors.

Language matters and shapes our culture. Casually assigning positive
connotations to white and negative connotations to black does not create
an inclusive culture. Blocklist and allowlist are also less ambiguious
terms for people unfamiliar with their concepts. Blocklist and allowlist
are a more accurate representation of what they're used for.