Revert changes to github buttons jsonp function and svg icon
We don't want to add github domain to our script src whitelist, which
is why jsonp funtion was rewritten. It was failing with api.github.com
being blocked by script-src csp.
the github svg icon was not showing up because data was not part of
our image-src. Adding it fixed the scp issue but the image was still
not showing up. Needed further debugging if we really want to use svg.

The icon connecting the star count and main icon is still broken.