Added MFA redirection/blocking for all routes that were previously blocked by a sign in, for users that need to have MFA enabled but have not yet enabled it or have it at a weak level.
Users who are mfa required but not enabled get redirected to the MFA setup page,
Users who are mfa required but at a weak level (UI Only) get redirected to the edit settings page to upgrade their MFA level.

Tests cover every route for the affected controllers. The affected controllers are:
api_keys, dashboard, email_confirmations, notifiers, ownership_calls, ownership_requests, profiles, settings

When users get redirected, they see a flash message with a link to the blog post. CSS for this blog post and a Ruby function
were added to enable this. A flash message type was added, notice_html, and anything with that type will render the html passed in.
The only thing using this feature is our fixed string flash warning, so no malicious code should be able to be run.
The CSS that was added makes the link black and instead of the normal blue.