Added a mfa_redirect_uri
Users who get redirected to set up or upgrade their MFA will now be redirected
back to the original page they were trying to access after they have completed the process.

The tests mainly target the GET paths because this is for quality of life,
and not security. Tests are still added to make sure that mfa_redirect_uri
is setup for all paths.

I couldn't use path helpers in the test files because of the context/should
setup, but using the actual paths with strings works fine.