Gem::SafeYAML is successfully upstreamed. Remove forbidden_yaml patch
This removes the patch on Gem::Specification and Gem::Package YAML
loading behavior. The patch has been upstream in rubygems in full for at
least 3 years. Since rubygems.org uses ruby 3.2.1, at this time, we no
longer need to confirm we are using a safe yaml correctly.

Mostly this was updated in this commit, but compatibility for older
versions was maintained, allowing the possibility of regression.
https://github.com/rubygems/rubygems/commit/510b1638ac9bba3ceb7a5d73135dafff9e5bab49