Re-add legacy field, mfa_seed in UserResource
Since previous audits would have captured the state of
mfa_seed field, removing the field from UserResource
would prevent that field from ever being rendered on the
audit show page.

This change adds the legacy field, mfa_seed, back in
so that we can continue to render it if an audit
includes this field.