Simply allow unsafe-inline for style-src is CSP
I noticed that GitHub.com does the same, so I interpret the risk
as very minimal. We don't interpret user input for display anywhere
on the site, and it removes this ugly script and fixes some other
minor bugs with inline svgs in tailwind.css

Also removes nonces from styles in CSP, which should be unused when
allowing unsafe-inline.