Fix the spec creating an API Key Role for Buildkite
The conditional in add_default_params is always try (it's a single =) so
statement.principal is set to the GitHub Actions principal
(https://token.actions.githubusercontent.com) every time the new API Key
Role form is loaded.

That field is hidden on the form so the user doesn't have a chance to
edit it, and after saving the created role has a provider of Buildkite
with an expected principal for GitHub Actions.

An alternative solution would be to remove the statement.principal
assignment completely. It's not required - when the form is submitted
the OIDC::ApiKeyRole#set_statement_principals callback will set the
correct principal for both GitHub Actionas *and* Buildkite.