New system test creating an API Key Role for Buildkite (#5434)
* Add a failing system test for creating an API Key Role for Buildkite

The form and POSTing to create works, but the resulting API Key Role has
a condition that expects a principal of "https://token.actions.githubusercontent.com"

* Fix the spec creating an API Key Role for Buildkite

When a gem has a source doe URI for GitHub, `add_default_params`
defaults the form to common settings for GitHub OIDC tokens.

That's a reasonable thing to do. However, that includes setting the
principal to a Github Actions shaped value. That value is rendered on
the form in hidden elements so the user doesn't have a chance to
edit it. After saving the created role has a provider of Buildkite
with an expected principal for GitHub Actions.

Its safe to remove the statement.principal assignment completely. It's
not required - when the form is submitted the
OIDC::ApiKeyRole#set_statement_principals callback will set the correct
principal for both GitHub Actions *and* Buildkite.