Fetch oldest authored_at correctly instead of oldest authored_at per page (#4460)
* Fetch oldest authored_at correctly instead of oldest authored_at per page

Fixes https://github.com/rubygems/rubygems.org/issues/4448

Previously, the "since date" was using oldest `version.authored_at` for a given
page. This change fixes it to get the oldest `version.authored_at` in
the database. Because `authored_at` depends on either `built_at` or
`created_at`, I fetch both and take the min. This should only be 2
additional queries of 1 record each.

I modified existing tests to check for the "versions since" text and
also have 2 pages of actual content.

* Add oldest_authored_at scope to Version

Replicate `Version#authored_at` logic in a SQL case statement and call
SQL `minimum` on that to get the oldest `authored_at`

Add tests for the model using similar structure as tests for `authored_at`

* Remove unnecessary in_time_zone and make oldest_authored_at a class method

Also now that `.oldest_authored_at` is a class method and not a scope,
moved it next to `#authored_at

---------

Co-authored-by: Albert Chae <albertchae@users.noreply.github.com>