🫗 Remove LaunchDarkly (#5780)
* Remove references to LaunchDarkly

This is step one in removing our reliance on LaunchDarkly, which has not
been used for a while now.

* Remove LaunchDarkly configuration

* Remove LaunchDarkly from Gemfile

* Add `zlib` to Gemfile to ensure it's present

We had previously had this listed as an indirect depenency of
LaunchDarkly, but that gem was removed. While this is part of the
standard library, it's not guaranteed to be compiled with Ruby. Adding
it to Gemfile will ensure that we have it when needed.