😅 Rectify LaunchDarkly state (#5782)
* Re-introduce StoreVersionContentsJob calling

This had been enabled in LaunchDarkly before the gem was removed.

* Remove previously-gated StatsD reporting

These had been configured to be disabled in LaunchDarkly, but the gem
was removed. Removing the calls now.

* Prevent job from running in test env

This is a fast followup re-introducing behavior that was configured in a
service, but removed.

* Remove assertions around dead code

These have been disabled in production for some time. Since we're not
using them, let's get rid of them. We can always add them back when
needed.

* Always run StoreVersionContentsJob

This had been enabled in production, but that was not reflected in
tests. I've stubbed StoreVersionContentsJob in VersionsSystemTest, which
I don't love. I'll look into related test setup and see what needs to be
done to actually exercise the code.